pyproject.toml
[tool.poetry]
name = "client-harmonic"
version = "0.1.0"
description = "The best harmonic client in the universe"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
black = "^24.3.0"
mypy = "^1.9.0"
loguru = "^0.7.2"
cryptography = "^42.0.5"
pycryptodome = "^3.20.0"
wxpython = "^4.2.1"
pyaudio = "^0.2.14"
librosa = "^0.10.2.post1"
keyboard = "^0.13.5"
noisereduce = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
